creates first 2 blocks

+++ src/game_logic/position.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct Position {
-    pub row: usize,
-    pub column: usize
+    pub row: i32,
+    pub column: i32
 }
